agents/calendar-agent/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.1
celery==5.3.4
redis==5.0.1
python-dateutil==2.8.2
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
msal==1.25.0
requests==2.31.0
python-dotenv==1.0.0
